PII/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
bcrypt==4.1.1
mysql-connector-python==8.2.0
pytesseract==0.3.13
Pillow==11.0.0
opencv-python==4.12.0.88
numpy==2.1.2
spacy==3.8.11
sentence-transformers==5.2.3
faiss-cpu==1.13.2
